image_analysis/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.2
blinker==1.8.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
debugpy==1.8.1
decorator==5.1.1
distro==1.9.0
Django==5.1.3
executing==2.0.1
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
ipykernel==6.29.2
ipython==8.21.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.7.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.2
openai==1.35.14
packaging==23.2
parso==0.8.3
pefile==2023.2.7
pillow==11.0.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
Pygments==2.17.2
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
python-dateutil==2.8.2
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.2
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
sentiment-analysis==0.1.5
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
